+++ common/src/main/java/net/mehvahdjukaar/sawmill/SawmillMenu.java
@@ -3,17 +3,14 @@ package net.mehvahdjukaar.sawmill;
 import com.google.common.collect.Lists;
 import net.minecraft.core.RegistryAccess;
 import net.minecraft.network.FriendlyByteBuf;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.Container;
 import net.minecraft.world.SimpleContainer;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.*;
-import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.crafting.RecipeHolder;
 import net.minecraft.world.level.Level;
 
 import java.util.List;
@@ -22,16 +19,16 @@ public class SawmillMenu extends AbstractContainerMenu {
     private final ContainerLevelAccess access;
     private final DataSlot selectedRecipeIndex;
     private final Level level;
-    private List<RecipeHolder<WoodcuttingRecipe>> recipes;
-    private ItemStack input;
-    long lastSoundTime;
-    final Slot inputSlot;
-    final Slot resultSlot;
-    Runnable slotUpdateListener;
     public final Container container;
-    final ResultContainer resultContainer;
+    private final Slot inputSlot;
+    private final Slot resultSlot;
 
-    private ResourceLocation lastSelectedRecipe = null;
+    private List<FilterableRecipe> recipes;
+    private ItemStack input;
+    private long lastSoundTime;
+    private final ResultContainer resultContainer;
+    private Runnable slotUpdateListener;
+    private FilterableRecipe lastSelectedRecipe = null;
 
     public SawmillMenu(int i, Inventory inventory, FriendlyByteBuf buf) {
         this(i, inventory, ContainerLevelAccess.NULL);
@@ -66,7 +63,8 @@ public class SawmillMenu extends AbstractContainerMenu {
             public void onTake(Player player, ItemStack stack) {
                 stack.onCraftedBy(player.level(), player, stack.getCount());
                 resultContainer.awardUsedRecipes(player, this.getRelevantItems());
-                ItemStack itemStack = inputSlot.remove(recipes.get(selectedRecipeIndex.get()).value().getInputCount());
+                ItemStack itemStack = inputSlot.remove(recipes.get(selectedRecipeIndex.get())
+                        .recipe().getInputCount());
                 if (!itemStack.isEmpty()) {
                     setupResultSlot();
                 }
@@ -105,7 +103,7 @@ public class SawmillMenu extends AbstractContainerMenu {
         return this.selectedRecipeIndex.get();
     }
 
-    public List<RecipeHolder<WoodcuttingRecipe>> getRecipes() {
+    public List<FilterableRecipe> getRecipes() {
         return this.recipes;
     }
 
@@ -120,17 +118,11 @@ public class SawmillMenu extends AbstractContainerMenu {
 
     @Override
     public boolean clickMenuButton(Player player, int id) {
-        if (this.isValidRecipeIndex(id)) {
+        if (this.isValidRecipeIndex(id) || id == -1) {
             this.selectedRecipeIndex.set(id);
             this.setupResultSlot();
         }
-
         return true;
-    }
-
-    public void clearResult() {
-        this.selectedRecipeIndex.set(-1);
-        this.setupResultSlot();
     }
 
     private boolean isValidRecipeIndex(int recipeIndex) {
@@ -142,8 +134,8 @@ public class SawmillMenu extends AbstractContainerMenu {
         ItemStack itemStack = this.inputSlot.getItem();
         ItemStack old = this.input;
         boolean sameStack = itemStack.is(old.getItem());
-        int maxConsumable = 5;
-        if (!sameStack || itemStack.getCount() < maxConsumable || old.getCount() < maxConsumable) {
+        int maxItemsThatCanBeConsumed = 5; //I made it the f up
+        if (!sameStack || itemStack.getCount() < maxItemsThatCanBeConsumed || old.getCount() < maxItemsThatCanBeConsumed) {
             this.input = itemStack.copy();
             this.setupRecipeList(container, itemStack);
         }
@@ -151,23 +143,26 @@ public class SawmillMenu extends AbstractContainerMenu {
     }
 
     private void setupRecipeList(Container container, ItemStack stack) {
-        this.recipes.clear();
         this.selectedRecipeIndex.set(-1);
 
         this.resultSlot.set(ItemStack.EMPTY);
         if (!stack.isEmpty()) {
-            this.recipes = this.level.getRecipeManager()
+            var matching = this.level.getRecipeManager()
                     .getRecipesFor(SawmillMod.WOODCUTTING_RECIPE.get(), container, this.level);
 
             //remove blacklisted
-            this.recipes.removeIf(r -> r.value().getResultItem(RegistryAccess.EMPTY).is(SawmillMod.BLACKLIST));
+            matching.removeIf(r -> r.getResultItem(RegistryAccess.EMPTY).is(SawmillMod.BLACKLIST));
 
-            RecipeSorter.sort(this.recipes, this.level);
+            RecipeSorter.sort(matching, this.level);
+
+            recipes = matching.stream().map(FilterableRecipe::of).toList();
 
             //preserve last clicked recipe on recipe change
             if (lastSelectedRecipe != null) {
-                this.recipes.stream().filter(h -> h.id().equals(lastSelectedRecipe)).findAny()
-                        .ifPresent(r -> this.selectedRecipeIndex.set(this.recipes.indexOf(r)));
+                int newInd = this.recipes.indexOf(lastSelectedRecipe);
+                if (newInd != -1) {
+                    this.selectedRecipeIndex.set(newInd);
+                }
             }
         }
 
@@ -176,11 +171,11 @@ public class SawmillMenu extends AbstractContainerMenu {
 
     void setupResultSlot() {
         if (!this.recipes.isEmpty() && this.isValidRecipeIndex(this.selectedRecipeIndex.get())) {
-            var recipe = this.recipes.get(this.selectedRecipeIndex.get());
-            this.lastSelectedRecipe = recipe.id();
-            ItemStack itemStack = recipe.value().assemble(this.container, this.level.registryAccess());
+            FilterableRecipe selected = this.recipes.get(this.selectedRecipeIndex.get());
+            this.lastSelectedRecipe = selected;
+            ItemStack itemStack = selected.recipe().assemble(this.container, this.level.registryAccess());
             if (itemStack.isItemEnabled(this.level.enabledFeatures())) {
-                this.resultContainer.setRecipeUsed(recipe);
+                this.resultContainer.setRecipeUsed(selected.recipe());
                 this.resultSlot.set(itemStack);
             } else {
                 this.resultSlot.set(ItemStack.EMPTY);
@@ -261,5 +256,6 @@ public class SawmillMenu extends AbstractContainerMenu {
             this.clearContainer(player, this.container);
         });
     }
+
 
 }