+++ common/src/main/java/net/mehvahdjukaar/sawmill/SawmillMod.java
@@ -68,7 +68,6 @@ public class SawmillMod {
         //TODO: add searchbar
         if (PlatHelper.getPhysicalSide().isClient()) {
             SawmillClient.init();
-            ClientConfigs.init();
         }
         CarpenterTrades.init();
         CommonConfigs.init();